Test case for where secondary arch in a newer version gets pulled in
When installing an older version of a package, the version gets
correctly installed for x86_64, but at the same time, a newer version of
the same package gets installed for i686 - it gets pulled in through
dependencies, but the older version for x86_64 should satisfy those as
well.

testcase.t show the errorneous, unexpected result.

The forcebest flag seems to make no difference here.